[pytorch] Add HashStore to c10d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28921

This implementation is quite similar to the HashStore in gloo -
an ephemeral in-process store with a lock and unordered_map<>
There are just small tweaks for the differences (e.g. add/check methods,
uint8_t vs char)

It's potentially a better choice for some cases than FileStore when we
don't need cross-process access, or care about the backing file.
ghstack-source-id: 92945875

Differential Revision: [D18233713](https://our.internmc.facebook.com/intern/diff/D18233713/)